Divide search range on Z axis
The search goes north and then west. The main loop is on the X axis.
Small optimisations are done so that the biome and swamp hut structure
don't have to be regenerated for a region that it was already generated
for in the previous iteration. However, this resets when moving up the Z
axis. It is therefore more reasonable the divide the search range on the
Z axis so that the X axis remains full width and the majority of the
scan can make use of the optimisation and reset less.